Implement object not found exception

+++ backend/Electric/src/main/java/com/example/electric/controller/AppointmentController.java
@@ -1,7 +1,7 @@
 package com.example.electric.controller;
 
 import com.example.electric.error.ErrorCode;
-import com.example.electric.exception.AppointmentNotFoundException;
+import com.example.electric.exception.ObjectNotFoundException;
 import com.example.electric.model.Appointment;
 import com.example.electric.service.AppointmentService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,7 +24,7 @@ public class AppointmentController {
     @GetMapping("/{appointmentId}")
     public Optional<Appointment> getAppointmentById(@PathVariable("appointmentId") long appointmentId) {
         if (!appointmentService.getAppointmentById(appointmentId).isPresent()) {
-            throw new AppointmentNotFoundException(ErrorCode.E1002);
+            throw new ObjectNotFoundException(ErrorCode.E1002);
         }
         return appointmentService.getAppointmentById(appointmentId);
     }
@@ -42,11 +42,17 @@ public class AppointmentController {
 
     @PutMapping("/{id}")
     public Appointment updateAppointment(@RequestBody Appointment updatedAppointment, @PathVariable("id") long id) {
+        if (!appointmentService.getAppointmentById(id).isPresent()) {
+            throw new ObjectNotFoundException(ErrorCode.E1002);
+        }
         return appointmentService.updateAppointment(updatedAppointment, id);
     }
 
     @DeleteMapping("/{id}")
     public void deleteAppointment(@PathVariable("id") long id) {
+        if (!appointmentService.getAppointmentById(id).isPresent()) {
+            throw new ObjectNotFoundException(ErrorCode.E1002);
+        }
         appointmentService.deleteAppointment(id);
     }
 }
