+++ backend/Electric/src/main/java/com/example/electric/controller/RecordController.java
@@ -1,5 +1,7 @@
 package com.example.electric.controller;
 
+import com.example.electric.error.ErrorCode;
+import com.example.electric.exception.ObjectNotFoundException;
 import com.example.electric.model.Record;
 import com.example.electric.service.RecordService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,6 +27,9 @@ public class RecordController {
 
     @GetMapping("/{id}")
     public Record getRecordById(@PathVariable Long id) {
+        if (recordService.getRecordById(id) == null) {
+            throw new ObjectNotFoundException(ErrorCode.E1002);
+        }
         return recordService.getRecordById(id);
     }
 
@@ -35,11 +40,17 @@ public class RecordController {
 
     @PutMapping("/{id}")
     public Record updateRecord(@PathVariable Long id, @RequestBody Record updatedRecord) {
+        if (recordService.getRecordById(id) == null) {
+            throw new ObjectNotFoundException(ErrorCode.E1002);
+        }
         return recordService.updateRecord(id, updatedRecord);
     }
 
     @DeleteMapping("/{id}")
     public void deleteRecord(@PathVariable Long id) {
+        if (recordService.getRecordById(id) == null) {
+            throw new ObjectNotFoundException(ErrorCode.E1002);
+        }
         recordService.deleteRecord(id);
     }
 }
